refactor: Better align REPL examples for comparison
