--- src/components/controllers/repl/examples/simple-counter.txt
@@ -1,16 +1,0 @@
-import { render } from 'preact';
-import { useState } from 'preact/hooks';
-
-function Counter() {
-	const [value, setValue] = useState(0);
-
-	return (
-		<>
-			<div>Counter: {value}</div>
-			<button onClick={() => setValue(value + 1)}>Increment</button>
-			<button onClick={() => setValue(value - 1)}>Decrement</button>
-		</>
-	);
-}
-
-render(<Counter />, document.getElementById('app'));
