+++ src/components/controllers/repl/examples/style.css
@@ -1,3 +1,15 @@
+.counter-container {
+	input,
+	button {
+		margin: 0.5rem;
+		text-align: center;
+	}
+
+	input {
+		width: 3rem;
+	}
+}
+
 .list-item {
 	padding: 1rem;
 	margin: 1rem;
